move functions to rocketchat folder
